frontend-app-api: refactor to add CoreRouter extension
